+++ node/_crypto/crypto_browserify/cipher_base_test.js
@@ -126,8 +126,7 @@ Deno.test("encodings", async function (t) {
     cipher.final("base64");
   });
   await t.step("handle long uft8 plaintexts", function () {
-    var txt =
-      "ふっかつ　あきる　すぶり　はやい　つける　まゆげ　たんさん　みんぞく　ねほりはほり　せまい　たいまつばな　ひはん";
+    var txt = "ふっかつ　あきる　すぶり　はやい　つける　まゆげ　たんさん　みんぞく　ねほりはほり　せまい　たいまつばな　ひはん";
 
     var cipher = new Cipher();
     var decipher = new Cipher();
